added the start of a new installer, added the latest version of dojo
git-svn-id: http://svn.psychdesktop.net/trunk@277 815afad1-63b8-44ff-8d19-411ecde00fa5

+++ desktop/dojo/dojo/dnd/source.js
@@ -34,12 +34,12 @@ function(node, params){
 	//	the rest of parameters are passed to the selector
 	if(!params){ params = {}; }
 	this.isSource = typeof params.isSource == "undefined" ? true : params.isSource;
-	var types = params.accept instanceof Array ? params.accept : ["text"];
+	var type = params.accept instanceof Array ? params.accept : ["text"];
 	this.accept = null;
-	if(types.length){
+	if(type.length){
 		this.accept = {};
-		for(var i = 0; i < types.length; ++i){
-			this.accept[types[i]] = 1;
+		for(var i = 0; i < type.length; ++i){
+			this.accept[type[i]] = 1;
 		}
 	}
 	this.horizontal = params.horizontal;
@@ -64,13 +64,66 @@ function(node, params){
 	}
 	// set up events
 	this.topics = [
-		dojo.subscribe("dndSourceOver", this, "onDndSourceOver"),
-		dojo.subscribe("dndStart",  this, "onDndStart"),
-		dojo.subscribe("dndDrop",   this, "onDndDrop"),
-		dojo.subscribe("dndCancel", this, "onDndCancel")
+		dojo.subscribe("/dnd/source/over", this, "onDndSourceOver"),
+		dojo.subscribe("/dnd/start",  this, "onDndStart"),
+		dojo.subscribe("/dnd/drop",   this, "onDndDrop"),
+		dojo.subscribe("/dnd/cancel", this, "onDndCancel")
 	];
 },
 {
+	// object attributes (for markup)
+	isSource: true,
+	horizontal: false,
+	copyOnly: false,
+	accept: ["text"],
+	
+	// methods
+	checkAcceptance: function(source, nodes){
+		// summary: checks, if the target can accept nodes from this source
+		// source: Object: the source which provides items
+		// nodes: Array: the list of transferred items
+		if(this == source){ return true; }
+		var accepted = true;
+		for(var i = 0; i < nodes.length; ++i){
+			var type = source.map[nodes[i].id].type;
+			if(type instanceof Array){
+				var flag = false;
+				for(var j = 0; j < type.length; ++j){
+					if(type[j] in this.accept){
+						flag = true;
+						break;
+					}
+				}
+				if(!flag){
+					accepted = false;
+					break;
+				}
+			}else{
+				accepted = false;
+			}
+			if(!accepted){ break; }
+		}
+		return accepted;	// Boolean
+	},
+	copyState: function(keyPressed){
+		// summary: Returns true, if we need to copy items, false to move.
+		//		It is separate to be overwritten dynamically, if needed.
+		// keyPressed: Boolean: the "copy" was pressed
+		return this.copyOnly || keyPressed;	// Boolean
+	},
+	destroy: function(){
+		// summary: prepares the object to be garbage-collected
+		dojo.dnd.Source.superclass.destroy.call(this);
+		dojo.forEach(this.topics, dojo.unsubscribe);
+		this.targetAnchor = null;
+	},
+
+	// markup methods
+	markupFactory: function(params, node){
+		params._skipStartup = true;
+		return new dojo.dnd.Source(node, params);
+	},
+
 	// mouse event processors
 	onMouseMove: function(e){
 		// summary: event processor for onmousemove
@@ -120,6 +173,7 @@ function(node, params){
 		this.mouseDown = false;
 		dojo.dnd.Source.superclass.onMouseUp.call(this, e);
 	},
+	
 	// topic event processors
 	onDndSourceOver: function(source){
 		// summary: topic event processor for ondndsourceover, called when detected a current source
@@ -156,23 +210,81 @@ function(node, params){
 		// copy: Boolean: copy items, if true, move items otherwise
 		do{ //break box
 			if(this.containerState != "Over"){ break; }
-			var oldCreator = this.nodeCreator;
-			if(this != source || copy){
-				this.selectNone();
-				this.nodeCreator = function(n){
-					return oldCreator(source.map[n.id].data);
-				};
+			var oldCreator = this._normalizedCreator;
+			if(this != source){
+				// transferring nodes from the source to the target
+				if(this.creator){
+					// use defined creator
+					this._normalizedCreator = function(node, hint){
+						return oldCreator.call(this, source.map[node.id].data, hint);
+					};
+				}else{
+					// we have no creator defined => move/clone nodes
+					if(copy){
+						// clone nodes
+						this._normalizedCreator = function(node, hint){
+							var t = source.map[node.id];
+							var n = node.cloneNode(true);
+							n.id = dojo.dnd.getUniqueId();
+							return {node: n, data: t.data, type: t.type};
+						};
+					}else{
+						// move nodes
+						this._normalizedCreator = function(node, hint){
+							var t = source.map[node.id];
+							delete source.map[node.id];
+							return {node: node, data: t.data, type: t.type};
+						};
+					}
+				}
 			}else{
+				// transferring nodes within the single source
 				if(this.current && this.current.id in this.selection){ break; }
-				this.nodeCreator = function(n){
-					var t = source.map[n.id]; return {node: n, data: t.data, types: t.types};
-				};
+				if(this.creator){
+					// use defined creator
+					if(copy){
+						// create new copies of data items
+						this._normalizedCreator = function(node, hint){
+							return oldCreator.call(this, source.map[node.id].data, hint);
+						};
+					}else{
+						// move nodes
+						this._normalizedCreator = function(node, hint){
+							var t = source.map[node.id];
+							return {node: node, data: t.data, type: t.type};
+						};
+					}
+				}else{
+					// we have no creator defined => move/clone nodes
+					if(copy){
+						// clone nodes
+						this._normalizedCreator = function(node, hint){
+							var t = source.map[node.id];
+							var n = node.cloneNode(true);
+							n.id = dojo.dnd.getUniqueId();
+							return {node: n, data: t.data, type: t.type};
+						};
+					}else{
+						// move nodes
+						this._normalizedCreator = function(node, hint){
+							var t = source.map[node.id];
+							return {node: node, data: t.data, type: t.type};
+						};
+					}
+				}
+			}
+			this._removeSelection();
+			if(this != source){
+				this._removeAnchor();
+			}
+			if(this != source && !copy && !this.creator){
+				source.selectNone();
 			}
 			this.insertNodes(true, nodes, this.before, this.current);
-			this.nodeCreator = oldCreator;
-			if(this != source && !copy){
+			if(this != source && !copy && this.creator){
 				source.deleteSelectedNodes();
 			}
+			this._normalizedCreator = oldCreator;
 		}while(false);
 		this.onDndCancel();
 	},
@@ -188,6 +300,7 @@ function(node, params){
 		this._changeState("Source", "");
 		this._changeState("Target", "");
 	},
+	
 	// utilities
 	onOverEvent: function(){
 		// summary: this function is called once, when mouse is over our container
@@ -198,46 +311,6 @@ function(node, params){
 		// summary: this function is called once, when mouse is out of our container
 		dojo.dnd.Source.superclass.onOutEvent.call(this);
 		dojo.dnd.manager().outSource(this);
-	},
-	// methods
-	destroy: function(){
-		// summary: prepares the object to be garbage-collected
-		dojo.dnd.Source.superclass.destroy.call(this);
-		dojo.forEach(this.topics, dojo.unsubscribe);
-		this.targetAnchor = null;
-	},
-	checkAcceptance: function(source, nodes){
-		// summary: checks, if the target can accept nodes from this source
-		// source: Object: the source which provides items
-		// nodes: Array: the list of transferred items
-		if(this == source){ return true; }
-		var accepted = true;
-		for(var i = 0; i < nodes.length; ++i){
-			var types = source.map[nodes[i].id].types;
-			if(types instanceof Array){
-				var flag = false;
-				for(var j = 0; j < types.length; ++j){
-					if(types[j] in this.accept){
-						flag = true;
-						break;
-					}
-				}
-				if(!flag){
-					accepted = false;
-					break;
-				}
-			}else{
-				accepted = false;
-			}
-			if(!accepted){ break; }
-		}
-		return accepted;	// Boolean
-	},
-	copyState: function(keyPressed){
-		// summary: Returns true, if we need to copy items, false to move.
-		//		It is separate to be overwritten dynamically, if needed.
-		// keyPressed: Boolean: the "copy" was pressed
-		return this.copyOnly || keyPressed;	// Boolean
 	},
 	_markTargetAnchor: function(before){
 		// summary: assigns a class to the current target anchor based on "before" status
@@ -273,6 +346,13 @@ function(node, params){
 	// summary: a constructor of the Target --- see the Source constructor for details
 	this.isSource = false;
 	dojo.removeClass(this.node, "dojoDndSource");
+},
+{
+	// markup methods
+	markupFactory: function(params, node){
+		params._skipStartup = true;
+		return new dojo.dnd.Target(node, params);
+	}
 });
 
 }
